infrastructure/data_pipeline_standalone/requirements.txt
# Core data processing
pandas>=1.5.0
polars>=0.19.0
numpy>=1.24.0

# Market data
yfinance>=0.2.28
requests>=2.31.0

# Azure connectivity
azure-storage-file-datalake>=12.12.0
pyodbc>=4.0.39

# Utilities
python-dotenv>=1.0.0
fsspec>=2023.6.0

# Optional: RL model inference (for Phase 7)
# Uncomment when models are trained
# ray[rllib]>=2.7.0
# torch>=2.0.0
# gymnasium>=0.29.0

